+++ R/Worker.R
@@ -6,7 +6,7 @@
 #' \code{\link[R6]{R6Class}} to create local and remote linux workers.
 #'
 #' @field nodename Host name. Set via constructor.
-#' @field ncpus Number of VPUs of worker. Set via constructor and defaults to the number of CPUs of the machine.
+#' @field ncpus Number of CPUs. Set via constructor and defaults to a heuristic which tries to detect the number of CPUs of the machine.
 #' @field max.load Maximum load average (of the last 5 min). Set via constructor and defaults to the number of CPUs of the machine.
 #' @field status Status of the worker; one of \dQuote{unknown}, \dQuote{available}, \dQuote{max.cpus} and \dQuote{max.load}.
 #' @section Methods: