+++ R/clearRegistry.R
@@ -6,7 +6,7 @@
 #' @family Registry
 #' @export
 clearRegistry = function(reg = getDefaultRegistry()) {
-  assertRegistry(reg, writeable = TRUE, running.ok = FALSE, sync = TRUE)
+  assertRegistry(reg, writeable = TRUE, running.ok = FALSE)
   info("Removing %i jobs ...", nrow(reg$status))
   reg$status = reg$status[FALSE]
   reg$defs = reg$defs[FALSE]