+++ R/clusterFunctions.R
@@ -2,7 +2,7 @@
 #'
 #' @description
 #' This is the constructor used to create \emph{custom} cluster functions.
-#' Note that some standard implementations for Torque, Slurm, LSF, SGE, etc. ship
+#' Note that some standard implementations for TORQUE, Slurm, LSF, SGE, etc. ship
 #' with the package.
 #'
 #' @param name [\code{character(1)}]\cr
@@ -47,7 +47,7 @@
 makeClusterFunctions = function(name, submitJob, killJob = NULL, listJobsQueued = NULL, listJobsRunning = NULL,
   array.var = NA_character_, store.job = FALSE, scheduler.delay = 0, hooks = list()) {
   assertList(hooks, types = "function", names = "unique")
-  assertSubset(names(hooks), batchtools$hooks$name)
+  assertSubset(names(hooks), unlist(batchtools$hooks, use.names = FALSE))
 
   setClasses(list(
       name = assertString(name, min.chars = 1L),
@@ -259,14 +259,14 @@ getBatchIds = function(reg, status = "all") {
   tab = data.table(batch.id = character(0L), status = character(0L))
   batch.id = NULL
 
-  if (status %in% c("all", "running") && !is.null(cf$listJobsRunning)) {
+  if (status %chin% c("all", "running") && !is.null(cf$listJobsRunning)) {
     "!DEBUG [getBatchIds]: Getting running Jobs"
     x = unique(cf$listJobsRunning(reg))
     if (length(x) > 0L)
       tab = rbind(tab, data.table(batch.id = x, status = "running"))
   }
 
-  if (status %in% c("all", "queued") && !is.null(cf$listJobsQueued)) {
+  if (status %chin% c("all", "queued") && !is.null(cf$listJobsQueued)) {
     "!DEBUG [getBatchIds]: Getting queued Jobs"
     x = unique(setdiff(cf$listJobsQueued(reg), tab$batch.id))
     if (length(x) > 0L)