+++ R/clusterFunctionsSGE.R
@@ -1,7 +1,7 @@
 #' @title ClusterFunctions for SGE Systems
 #'
 #' @description
-#' Cluster functions for Univia Grid Engine / Oracle Grid Engine /
+#' Cluster functions for Univa Grid Engine / Oracle Grid Engine /
 #' Sun Grid Engine (\url{http://www.univa.com/}).
 #'
 #' Job files are created based on the brew template \code{template}. This