+++ R/execJob.R
@@ -31,7 +31,7 @@ execJob.JobCollection = function(job) {
 
 #' @export
 execJob.Job = function(job) {
-  if (".job" %in% names(formals(job$fun))) {
+  if (".job" %chin% names(formals(job$fun))) {
     with_seed(job$seed, do.call(job$fun, c(job$pars, list(.job = job)), envir = .GlobalEnv))
   } else {
     with_seed(job$seed, do.call(job$fun, job$pars, envir = .GlobalEnv))