+++ R/zzz.R
@@ -24,7 +24,7 @@
 #' @importFrom brew brew
 #' @importFrom progress progress_bar
 #' @importFrom rappdirs user_config_dir
-#' @importFrom stats runif predict
+#' @importFrom stats runif predict pexp
 "_PACKAGE"
 
 #' @title Deprecated function in the batchtools package
@@ -39,9 +39,9 @@ NULL
 
 batchtools = new.env(parent = emptyenv())
 batchtools$debug = FALSE
-batchtools$hooks = data.table(
-  name =   c("pre.sync", "post.sync", "pre.do.collection", "post.do.collection", "pre.submit.job", "post.submit.job", "pre.submit", "post.submit", "pre.kill", "post.kill"),
-  remote = c(FALSE, FALSE, TRUE, TRUE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE)
+batchtools$hooks = list(
+  remote = c("pre.do.collection", "post.do.collection"),
+  local  = c("pre.sync", "post.sync", "pre.submit.job", "post.submit.job", "pre.submit", "post.submit", "pre.kill", "post.kill")
 )
 
 .onLoad = function(libname, pkgname) { # nocov start